403 users from messing around with other user's sheets

+++ treatment_sheets/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render
 from django.shortcuts import redirect
 from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 from treatment_sheets.forms import TxSheetForm, TxItemForm
 from treatment_sheets.models import TxSheet, TxItem
@@ -19,14 +20,18 @@ def view_treatment_sheet(request, sheet_id):
     sheet = TxSheet.objects.get(pk=sheet_id)
     form = TxItemForm()
 
-    if request.method == 'POST':
+    if request.method == 'POST' and request.user == sheet.owner:
         form = TxItemForm(data=request.POST)
 
         if form.is_valid():
             tx_sheet = form.save(sheet=sheet)
             return redirect(tx_sheet)
 
-    return render(request, 'tx_sheet/tx_sheet_view.html', {'navbar': 'tx_sheet', 'sheet': sheet, 'form': form})
+    if request.user == sheet.owner:
+        return render(request, 'tx_sheet/tx_sheet_view.html', {'navbar': 'tx_sheet', 'sheet': sheet, 'form': form})
+
+    else:
+        raise PermissionDenied
 
 
 @login_required()
@@ -50,5 +55,10 @@ def new_tx_sheet(request):
 @login_required()
 def del_item_tx_sheet(request, sheet_id, item_id):
     tx_sheet = TxSheet.objects.get(id=sheet_id)
-    TxItem.objects.get(id=item_id).delete()
-    return redirect(tx_sheet)
+
+    if request.user == tx_sheet.owner:
+        TxItem.objects.get(id=item_id).delete()
+        return redirect(tx_sheet)
+
+    else:
+        raise PermissionDenied
